(issue #359) Created progress bar .c and .h files.
This commit creates ST_progressBar.c and ST_progressBar.h. These files make the progress bar stand-alone (except for a reference to SuperGlobals). Here are summaries for all of the files I modifed:

ST_grid.c: I removed the definitions for logProgress and calculateProgress, as well as the declaration of calculateProgress.

ST_grid.h: I removed the Status enumerator as well as the declaration for logProgress.

ST_progressBar.h: I added the Status enumerator and the declaration for logProgress.

ST_progressBar.c: I added the definitions for logProgress and calculateProgress (renamed _calculateProgress). I also add the declaration for _calculateProgress. Last, I added two new functions called _calculateInitializationProgress and _calculateSimulationProgress. These two functions make the module easier to understand and expand.

ST_main.c: Added an include statement for the progress bar module, then implemented the progress bar. I also removed the "."s that the program was printing in place of a progress bar.

ST_initialization.c: Added an include statement for the progress bar module.

makefile: Added rule for ST_progressBar.c
